Moving zod validation out.
